Use MockedTestProvider
