--- client/shared/src/testing/apollo.ts
@@ -1,7 +1,0 @@
-import { act } from '@testing-library/react'
-
-/*
- * Wait one tick to load the next response from Apollo
- * https://www.apollographql.com/docs/react/development-testing/testing/#testing-the-success-state
- */
-export const waitForNextApolloResponse = (): Promise<void> => act(() => new Promise(resolve => setTimeout(resolve, 0)))
